requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.7
Django==5.1.3
djangorestframework==3.15.2
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
sqlparse==0.5.2
Werkzeug==3.1.3
